separeated user from auth
